+++ CHANGELOG.md
@@ -11,6 +11,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 - Support `OTEL_EXPORTER_ZIPKIN_ENDPOINT` env to specify zipkin collector endpoint (#2490)
+- Metrics Exponential Histogram support: Mapping functions have been made available
+  in `sdk/metric/aggregator/exponential/mapping` for other OpenTelemetry projects to take
+  dependencies on. (#X)
 
 ### Changed